requirements.txt
Pillow
pytesseract
flask >= 3.0.3